feat(web): extract agents stats helpers into agents-stats-lib
Move agent report assembly into agents-stats-lib with expanded Vitest coverage.

Preserve the public @/lib/agents-stats import surface.

Closes #4302
